Rename to _rdf_utils and add to CMakeLists in current PyROOT
