+++ bindings/pyroot_experimental/PyROOT/python/ROOT/pythonization/_rdataframe.py
@@ -39,7 +39,7 @@ def RDataFrameAsNumpy(df, columns=None, exclude=None):
     # Import numpy and numpy.array derived class lazily
     try:
         import numpy
-        from ROOT.pythonization._ndarray import ndarray
+        from ROOT.pythonization._rdf_utils import ndarray
     except:
         raise ImportError("Failed to import numpy during call of RDataFrame.AsNumpy.")
